+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add `Model#present?` and `Cell#present?`
 * Allow `Cell#save` to be called even without making any changes
 * Add `Model#fetch` and `Cell#fetch`
 * Add `Model#reload` and `Cell#reload`